% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcBurst.R
\name{CalcBurst}
\alias{CalcBurst}
\title{Create a burst vector based on relocations of one or multiple animals given a max time between intervals}
\usage{
CalcBurst(
  data = data,
  id = TRUE,
  id_name = "id",
  date_name = "date",
  Tmax = 10800
)
}
\arguments{
\item{data}{A dataframe or sf POINT dataframe with a posix column and an animal id column (optional)}

\item{id}{Logical. If TRUE (the default), look for and use an animal ID column to create bursts. If FALSE, assumes a single animal in the dataset.}

\item{id_name}{A character specifying the name of the column representing animal ID.}

\item{date_name}{A character specifying the name of the column representing date and time stamps of the locations.}

\item{Tmax}{An integer representing the max time (in seconds) between two relocations where those two relocations should be considered a connected step.}
}
\value{
Returns a vector with integers of the same value denoting the same burst. If the integers change their value, a new burst has begun.
}
\description{
It is based on two points connected in time, where row i is connected with row i+1. Written and updated by Jerod Merkle. Last updated January 2021.
}
\examples{
# none

}
