% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcMovParams.R
\name{CalcMovParams}
\alias{CalcMovParams}
\title{Add movement parameters to a sf point dataframe based on x/y values (in meters), time( asPOSIXct), and a vector of bursts.}
\usage{
CalcMovParams(
  data = data,
  id_name = "id",
  date_name = "date",
  burst = data$burst
)
}
\arguments{
\item{data}{A dataframe or sf POINT dataframe with a posix column and an animal id column}

\item{id_name}{A character specifying the name of the column representing animal ID.}

\item{date_name}{A character specifying the name of the column representing date and time stamps of the locations.}

\item{burst}{A vector of bursts, based on the CalcBurst function.}
}
\value{
Returns the same data.frame entered with the addition of the following columns: dist (distance between steps in meters), dt (time elapsed bewteen steps in seconds), speed (meters/second), abs.angle (angle moved relative no north, in degrees), rel.angle (angle moved relative to the direction of the previous step, in degrees), and StepFlag (logical, denoting whether a step is connected and has all movement params calculated)
}
\description{
movement params will only be calculated for sequential relocations denoted by the burst vector. Written by Jerod Merkle. Last updated January 2021.
}
\examples{
#order database first
data <- data[order(data$id, data$date),]
# example with 3 hr relocation data and multiple ids. Setting Tmax to 7 hours, and thus two points are connected even if a missed fix in between them (i.e., a 6 hr step).
data$burst <- CalcBurst(data=data, id = TRUE, id_name="id", date_name="date", Tmax = 3600*7)
data <- CalcMovParams(data=data, id_name = "id", date_name = "date", burst=data$burst)
}
