% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindProblemPts.R
\name{FindProblemPts}
\alias{FindProblemPts}
\title{Identify problem points.}
\usage{
FindProblemPts(
  data = data,
  date_name = "date",
  burst = data$burst,
  id_name = "id",
  speedlim = 3
)
}
\arguments{
\item{data}{A dataframe or sf POINT dataframe with a posix column and an animal id column (optional)}

\item{date_name}{A character specifying the name of the column representing date and time stamps of the locations.}

\item{burst}{A vector of bursts, based on the CalcBurst function.}

\item{id_name}{A character specifying the name of the column representing animal ID.}

\item{speedlim}{A character specifying the name of the column representing date and time stamps of the locations.}
}
\value{
Returns a logical vector indicating whether each row of dat is a problem point or not.
}
\description{
A problem point is defined as a point where there is a high speed coming into the point and going away from the point. Written and updated by Jerod Merkle. Last updated January 2021.
}
\examples{
#To come

}
